split the logger and removed PSR dependency
