+++ src/Logger/LoggerInterface.php
@@ -1,0 +1,142 @@
+<?php
+
+/**
+ * This file is part of the Phalcon Framework.
+ *
+ * (c) Phalcon Team <team@phalcon.io>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Phalcon\Logger;
+
+use Phalcon\Logger\Adapter\AdapterInterface;
+
+/**
+ * Interface for Phalcon based logger objects.
+ */
+interface LoggerInterface
+{
+    /**
+     * Action must be taken immediately.
+     *
+     * Example: Entire website down, database unavailable, etc. This should
+     * trigger the SMS alerts and wake you up.
+     *
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     */
+    public function alert(string $message, array $context = array()): void;
+
+    /**
+     * Critical conditions.
+     *
+     * Example: Application component unavailable, unexpected exception.
+     *
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     */
+    public function critical(string $message, array $context = array()): void;
+
+    /**
+     * Detailed debug information.
+     *
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     */
+    public function debug(string $message, array $context = array()): void;
+
+    /**
+     * System is unusable.
+     *
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     */
+    public function emergency(string $message, array $context = array()): void;
+
+    /**
+     * Runtime errors that do not require immediate action but should typically
+     * be logged and monitored.
+     *
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     */
+    public function error(string $message, array $context = array()): void;
+
+    /**
+     * Returns an adapter from the stack
+     *
+     * @param string $name The name of the adapter
+     *
+     * @return AdapterInterface
+     * @throws Exception
+     */
+    public function getAdapter(string $name): AdapterInterface;
+
+    /**
+     * Returns the adapter stack array
+     *
+     * @return AdapterInterface[]
+     */
+    public function getAdapters(): array;
+
+    /**
+     * Returns the log level
+     */
+    public function getLogLevel(): int;
+
+    /**
+     * Returns the name of the logger
+     */
+    public function getName(): string;
+
+    /**
+     * Interesting events.
+     *
+     * Example: User logs in, SQL logs.
+     *
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     */
+    public function info(string $message, array $context = array()): void;
+
+    /**
+     * Logs with an arbitrary level.
+     *
+     * @param mixed  $level
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     *
+     * @throws Exception
+     */
+    public function log($level, string $message, array $context = array());
+
+    /**
+     * Normal but significant events.
+     *
+     * @param string $message
+     * @param array  $context
+     *
+     * @return void
+     */
+    public function notice(string $message, array $context = array()): void;
+
+    public function warning(string $message, array $context = array()): void;
+}
